refactor: real type changed to double precision

+++ src/main/java/type/primitive/floating/Real.java
@@ -2,27 +2,20 @@ package type.primitive.floating;
 
 public class Real extends FloatBaseType{
 
-    public static final Float MAX_VALUE = Float.MAX_VALUE;
-    public static final Float MIN_VALUE = Float.MIN_VALUE;
+    public static final Double MAX_VALUE = Double.MAX_VALUE;
+    public static final Double MIN_VALUE = Double.MIN_VALUE;
     //private float value;
 
     public Real() {
         super("real64");
     }
 
-    public Real(float value) {
+    public Real(double value) {
         this.value = value;
     }
 
     public Real(String value) {
-        this.value = Float.parseFloat(value);
+        this.value = Double.parseDouble(value);
     }
 
-    //public float getValue() {
-    //    return value;
-    //}
-    //
-    //public void setValue(float value) {
-    //    this.value = value;
-    //}
 }
